requirement.txt
sanic
sanic-restplus
sanic_cors
sanic-plugins-framework

tortoise-orm
aiosqlite